Copied unit tests from search in rotated sorted array to be used in minimum in rotated sorted array
